requirements.txt
Flask
chatterbot
chatterbot-corpus
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
gunicorn #for deployment - production-grade WSGI server
psycopg2-binary
flask-sqlalchemy
flask-migrate
python-dotenv
Flask-Login
PyJWT
Werkzeug
authlib
requests
openai